requirements.txt
fastapi
uvicorn[standard]
pymongo
kafka-python
openai<1.0.0
python-dotenv
requests